Use correct namespace and save audio to save_dir
There was a bug where the ViddlRb namespace wasn't used by downloader.rb
causing the audio extraction to fail. Also, the audio extractor did not
save the audio file to the folder specified by the save_dir option.

It would probably be good to push a new version considering audio
extraction was broken.