Remove grep in favor of regex
grep is not included with Windows so this makes it easier to run the
tests on Windows. I got it working before with a Windows grep version,
but I could not get it working when I tried later on a different
machine. So I think it's better if we just use regular expressions to
extract the HTTP code.